refactor: view폴더 제거
