+++ src/index.js
@@ -1,8 +1,8 @@
 import '../style.css';
 import WebController from './Controller/WebController';
 
-import Dropdown from './view/components/Dropdown';
-import RestaurantCards from './view/components/RestaurantCards';
+import Dropdown from './components/Dropdown';
+import RestaurantCards from './components/RestaurantCards';
 
 window.customElements.define('drop-down', Dropdown, { extends: 'select' });
 window.customElements.define('restaurant-cards', RestaurantCards, { extends: 'ul' });
